Correction of dependency specifications for two common make rules
It was noticed that make pattern rules did not work as it was “hoped”
when two “patterns” were specified in the prerequisites list for
OCaml source files and their corresponding compiled interface descriptions.

This issue can be clarified further by the bug report
“Checking search retries for implicit make rules” (from 2017-06-25).
https://savannah.gnu.org/bugs/?51311

It seems that a replacement with static pattern rules was more appropriate
in this use case.
https://www.gnu.org/software/make/manual/html_node/Static-Usage.html

But they have got an other software behaviour with the consequence
that messages like the following can occur then usually.

“…
…/rule_pair.make:13: warning: overriding recipe for target 'ast_ctl.cmo'
…/common_rules.make:11: warning: ignoring old recipe for target 'ast_ctl.cmo'
…”


Fixes: 9597598779f3604272621f6d42ead79eaafad2a8 ("Bug 104: Improve some dependency specifications in make rules")

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>